+++ src/auto-evo/mutation_strategy/IMutationStrategy.cs
@@ -27,6 +27,6 @@ public interface IMutationStrategy<T>
     ///   List of mutated species, null if no possible mutations are found (some strategies may return an empty list
     ///   instead in this case)
     /// </returns>
-    public List<Tuple<T, float>>? MutationsOf(T baseSpecies, float mp, bool lawk, Random random,
+    public List<Tuple<T, double>>? MutationsOf(T baseSpecies, double mp, bool lawk, Random random,
         BiomeConditions biomeToConsider);
 }
